fix: Remove from history redirection
